net: reset session phase when a client joins another game
The same NetworkManager instance is reused when a player returns from a match and
joins or hosts again, so connect() and startServer() now clear the in-progress phase,
the simulation seed and any half-finished mod transfer. Without that reset the newly
strict phase checks would refuse the whole lobby handshake of the second game.

Also tightens two comparisons that mix a macro constant with an unsigned length.

+++ include/Network/ENetPacketIStream.h
@@ -83,7 +83,8 @@ public:
 
         // Subtraction form: currentPos is never allowed past dataLength, so this cannot wrap.
         // The additive form (currentPos + length) overflows on 32-bit size_t targets (wasm32).
-        if(length > ENETPACKET_MAX_STRING_LENGTH || length > getRemainingLength()) {
+        if(length > static_cast<Uint32>(ENETPACKET_MAX_STRING_LENGTH)
+           || static_cast<size_t>(length) > getRemainingLength()) {
             THROW(InputStream::eof, "ENetPacketIStream::readString(): End-of-File reached!");
         }
